+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+Additions:
+- `lsp-document-symbols` now renders symbols in a tree.
+- `rust-analyzer-run-test` to run test at cursor.
 Fixes:
 - Completion snippets (accessed via `lsp-snippets-select-next-placeholders`) can now be nested, making it possible to cycle through the arguments of nested function calls.
 - `completionItem.additionalTextEdits` are now applied also when the server does not support `completionItem/resolve`.